chore: update iced to 0.8

+++ src/main.rs
@@ -32,7 +32,7 @@ struct Cli {
 
     /// Font size
     #[clap(short, long)]
-    font_size: Option<u16>,
+    font_size: Option<f32>,
 }
 
 fn main() -> iced::Result {
@@ -62,7 +62,7 @@ fn main() -> iced::Result {
     info!("Path to XMonad config file: {}", &user_config.xmonad_config);
 
     let mut settings = Settings {
-        default_text_size: 22,
+        default_text_size: 22.0,
         default_font: Some(include_bytes!("../assets/fonts/Roboto-Regular.ttf")),
         ..Settings::with_flags(AppConfig::from(user_config))
     };
